requirements.txt
fastapi
pydantic
uvicorn
pymongo
passlib
google-genai
pdfplumber
google.generativeai
